+++ bin/miden-cli/tests/cli.rs
@@ -9,7 +9,7 @@ use std::{
 
 use assert_cmd::Command;
 use miden_client::{
-    self, Client, Felt,
+    self, Client, ExecutionOptions, Felt,
     account::{AccountId, AccountStorageMode},
     crypto::{FeltRng, RpoRandomCoin},
     note::{
@@ -28,6 +28,7 @@ use miden_client::{
     utils::Serializable,
 };
 use miden_client_cli::CliKeyStore;
+use miden_objects::{MAX_TX_EXECUTION_CYCLES, MIN_TX_EXECUTION_CYCLES};
 use predicates::str::contains;
 use rand::Rng;
 use toml::Table;
@@ -53,7 +54,7 @@ use uuid::Uuid;
 // ================================================================================================
 
 #[test]
-fn test_init_without_params() {
+fn init_without_params() {
     let temp_dir = init_cli().1;
 
     // Trying to init twice should result in an error
@@ -63,7 +64,7 @@ fn test_init_without_params() {
 }
 
 #[test]
-fn test_init_with_params() {
+fn init_with_params() {
     let store_path = create_test_store_path();
     let temp_dir = init_cli_with_store_path("devnet", &store_path);
 
@@ -88,7 +89,7 @@ fn test_init_with_params() {
 
 /// This test tries to run a mint TX using the CLI for an account that isn't tracked.
 #[tokio::test]
-async fn test_mint_with_untracked_account() {
+async fn mint_with_untracked_account() {
     let temp_dir = init_cli().1;
 
     // Create faucet account
@@ -109,7 +110,7 @@ async fn test_mint_with_untracked_account() {
 
 /// This test tries to run a mint TX using the CLI for an account that isn't tracked.
 #[tokio::test]
-async fn test_token_symbol_mapping() {
+async fn token_symbol_mapping() {
     let (store_path, temp_dir) = init_cli();
 
     // Create faucet account
@@ -173,7 +174,7 @@ const GENESIS_ACCOUNTS_FILENAMES: [&str; 1] = ["account.mac"];
 // 4. Runs a mint tx and syncs until the transaction and note are committed
 #[tokio::test]
 #[ignore = "import genesis test gets ignored by default so integration tests can be ran with dockerized and remote nodes where we might not have the genesis data"]
-async fn test_import_genesis_accounts_can_be_used_for_transactions() {
+async fn import_genesis_accounts_can_be_used_for_transactions() {
     let (store_path, temp_dir) = init_cli();
 
     for genesis_account_filename in GENESIS_ACCOUNTS_FILENAMES {
@@ -234,7 +235,7 @@ async fn test_import_genesis_accounts_can_be_used_for_transactions() {
 // 3. On client A runs a mint transaction, and exports the output note
 // 4. On client B imports the note and consumes it
 #[tokio::test]
-async fn test_cli_export_import_note() {
+async fn cli_export_import_note() {
     const NOTE_FILENAME: &str = "test_note.mno";
 
     let temp_dir_1 = init_cli().1;
@@ -299,7 +300,7 @@ async fn test_cli_export_import_note() {
 }
 
 #[tokio::test]
-async fn test_cli_export_import_account() {
+async fn cli_export_import_account() {
     const FAUCET_FILENAME: &str = "test_faucet.mac";
     const WALLET_FILENAME: &str = "test_wallet.wal";
 
@@ -364,7 +365,7 @@ async fn test_cli_export_import_account() {
 }
 
 #[test]
-fn test_cli_empty_commands() {
+fn cli_empty_commands() {
     let temp_dir = init_cli().1;
 
     let mut create_faucet_cmd = Command::cargo_bin("miden-client").unwrap();
@@ -386,7 +387,7 @@ fn test_cli_empty_commands() {
 }
 
 #[tokio::test]
-async fn test_consume_unauthenticated_note() {
+async fn consume_unauthenticated_note() {
     let temp_dir = init_cli().1;
 
     // Create wallet account
@@ -408,7 +409,7 @@ async fn test_consume_unauthenticated_note() {
 // ================================================================================================
 
 #[tokio::test]
-async fn test_init_with_devnet() {
+async fn init_with_devnet() {
     let temp_dir = init_cli_with_store_path("devnet", &create_test_store_path());
 
     // Check in the config file that the network is devnet
@@ -422,7 +423,7 @@ async fn test_init_with_devnet() {
 }
 
 #[tokio::test]
-async fn test_init_with_testnet() {
+async fn init_with_testnet() {
     let temp_dir = init_cli_with_store_path("testnet", &create_test_store_path());
 
     // Check in the config file that the network is testnet
@@ -771,7 +772,13 @@ async fn create_rust_client_with_store_path(store_path: &Path) -> (TestClient, C
             rng,
             store,
             std::sync::Arc::new(keystore.clone()),
-            true,
+            ExecutionOptions::new(
+                Some(MAX_TX_EXECUTION_CYCLES),
+                MIN_TX_EXECUTION_CYCLES,
+                false,
+                true,
+            )
+            .expect("Default executor's options should always be valid"),
             None,
             None,
         ),
@@ -791,7 +798,7 @@ fn assert_command_fails_but_does_not_panic(command: &mut Command) {
 // ================================================================================================
 
 #[test]
-fn test_exec_parse() {
+fn exec_parse() {
     let failure_script =
         fs::canonicalize("tests/files/test_cli_advice_inputs_expect_failure.masm").unwrap();
     let success_script =