+++ crates/rust-client/src/transaction/mod.rs
@@ -71,16 +71,16 @@ use alloc::{
 use core::fmt::{self};
 
 use miden_objects::{
-    AssetError, Digest, Felt, Word,
+    AssetError, Digest, Felt,
     account::{Account, AccountCode, AccountDelta, AccountId},
     assembly::DefaultSourceManager,
     asset::{Asset, NonFungibleAsset},
     block::BlockNumber,
-    note::{Note, NoteDetails, NoteId, NoteTag},
+    note::{Note, NoteDetails, NoteId, NoteRecipient, NoteTag},
     transaction::{AccountInputs, TransactionArgs},
 };
 use miden_tx::{
-    NoteAccountExecution, NoteConsumptionChecker,
+    NoteAccountExecution, NoteConsumptionChecker, TransactionExecutor,
     utils::{ByteReader, ByteWriter, Deserializable, DeserializationError, Serializable},
 };
 use tracing::info;
@@ -92,7 +92,7 @@ use crate::{
     rpc::domain::account::AccountProof,
     store::{
         InputNoteRecord, InputNoteState, NoteFilter, OutputNoteRecord, StoreError,
-        TransactionFilter, input_note_states::ExpectedNoteState,
+        TransactionFilter, data_store::ClientDataStore, input_note_states::ExpectedNoteState,
     },
     sync::NoteTagRecord,
 };
@@ -142,7 +142,7 @@ impl TransactionResult {
     /// [`TransactionResult`].
     pub async fn new(
         transaction: ExecutedTransaction,
-        note_screener: NoteScreener<'_>,
+        note_screener: NoteScreener,
         partial_notes: Vec<(NoteDetails, NoteTag)>,
         current_block_num: BlockNumber,
         current_timestamp: Option<u64>,
@@ -527,8 +527,8 @@ impl Client {
     ///
     /// # Errors
     ///
-    /// - Returns [`ClientError::MissingOutputNotes`] if the [`TransactionRequest`] ouput notes are
-    ///   not a subset of executor's output notes.
+    /// - Returns [`ClientError::MissingOutputRecipients`] if the [`TransactionRequest`] ouput notes
+    ///   are not a subset of executor's output notes.
     /// - Returns a [`ClientError::TransactionExecutorError`] if the execution fails.
     /// - Returns a [`ClientError::TransactionRequestError`] if the request is invalid.
     pub async fn new_transaction(
@@ -589,15 +589,15 @@ impl Client {
             InputNotes::new(input_notes).map_err(ClientError::TransactionInputError)?
         };
 
-        let output_notes: Vec<Note> =
-            transaction_request.expected_output_notes().cloned().collect();
+        let output_recipients =
+            transaction_request.expected_output_recipients().cloned().collect::<Vec<_>>();
 
         let future_notes: Vec<(NoteDetails, NoteTag)> =
             transaction_request.expected_future_notes().cloned().collect();
 
         let tx_script = transaction_request.build_transaction_script(
             &self.get_account_interface(account_id).await?,
-            self.in_debug_mode,
+            self.in_debug_mode(),
         )?;
 
         let foreign_accounts = transaction_request.foreign_accounts().clone();
@@ -607,6 +607,11 @@ impl Client {
             self.retrieve_foreign_account_inputs(foreign_accounts).await?;
 
         let ignore_invalid_notes = transaction_request.ignore_invalid_input_notes();
+
+        let data_store = ClientDataStore::new(self.store.clone());
+        for fpi_account in &foreign_account_inputs {
+            data_store.mast_store().load_account_code(fpi_account.code());
+        }
 
         let tx_args = transaction_request.into_transaction_args(tx_script, foreign_account_inputs);
 
@@ -623,7 +628,7 @@ impl Client {
             .await?
             .ok_or(ClientError::AccountDataNotFound(account_id))?;
         let account: Account = account_record.into();
-        self.mast_store.load_account_code(account.code());
+        data_store.mast_store().load_account_code(account.code());
 
         if ignore_invalid_notes {
             // Remove invalid notes
@@ -632,7 +637,7 @@ impl Client {
 
         // Execute the transaction and get the witness
         let executed_transaction = self
-            .tx_executor
+            .build_executor(&data_store)?
             .execute_transaction(
                 account_id,
                 block_num,
@@ -642,31 +647,11 @@ impl Client {
             )
             .await?;
 
-        // Check that the expected output notes matches the transaction outcome.
-        // We compare authentication commitments where possible since that involves note IDs +
-        // metadata (as opposed to just note ID which remains the same regardless of
-        // metadata) We also do the check for partial output notes
-
-        let tx_note_auth_commitments: BTreeSet<Digest> =
-            notes_from_output(executed_transaction.output_notes())
-                .map(Note::commitment)
-                .collect();
-
-        let missing_note_ids: Vec<NoteId> = output_notes
-            .iter()
-            .filter_map(|n| (!tx_note_auth_commitments.contains(&n.commitment())).then_some(n.id()))
-            .collect();
-
-        if !missing_note_ids.is_empty() {
-            return Err(ClientError::MissingOutputNotes(missing_note_ids));
-        }
-
-        let screener =
-            NoteScreener::new(self.store.clone(), &self.tx_executor, self.mast_store.clone());
+        validate_executed_transaction(&executed_transaction, &output_recipients)?;
 
         TransactionResult::new(
             executed_transaction,
-            screener,
+            NoteScreener::new(self.store.clone(), self.authenticator.clone()),
             future_notes,
             self.get_sync_height().await?,
             self.store.get_current_timestamp(),
@@ -814,17 +799,9 @@ impl Client {
     }
 
     /// Compiles the provided transaction script source and inputs into a [`TransactionScript`].
-    pub fn compile_tx_script<T>(
-        &self,
-        inputs: T,
-        program: &str,
-    ) -> Result<TransactionScript, ClientError>
-    where
-        T: IntoIterator<Item = (Word, Vec<Felt>)>,
-    {
-        let assembler = TransactionKernel::assembler().with_debug_mode(self.in_debug_mode);
-        TransactionScript::compile(program, inputs, assembler)
-            .map_err(ClientError::TransactionScriptError)
+    pub fn compile_tx_script(&self, program: &str) -> Result<TransactionScript, ClientError> {
+        let assembler = TransactionKernel::assembler().with_debug_mode(self.in_debug_mode());
+        TransactionScript::compile(program, assembler).map_err(ClientError::TransactionScriptError)
     }
 
     // HELPERS
@@ -839,15 +816,17 @@ impl Client {
     ) -> (BTreeMap<AccountId, u64>, BTreeSet<NonFungibleAsset>) {
         // Get own notes assets
         let mut own_notes_assets = match transaction_request.script_template() {
-            Some(TransactionScriptTemplate::SendNotes(notes)) => {
-                notes.iter().map(|note| (note.id(), note.assets())).collect::<BTreeMap<_, _>>()
-            },
+            Some(TransactionScriptTemplate::SendNotes(notes)) => notes
+                .iter()
+                .map(|note| (note.id(), note.assets().clone()))
+                .collect::<BTreeMap<_, _>>(),
             _ => BTreeMap::default(),
         };
         // Get transaction output notes assets
         let mut output_notes_assets = transaction_request
-            .expected_output_notes()
-            .map(|note| (note.id(), note.assets()))
+            .expected_output_own_notes()
+            .into_iter()
+            .map(|note| (note.id(), note.assets().clone()))
             .collect::<BTreeMap<_, _>>();
 
         // Merge with own notes assets and delete duplicates
@@ -995,7 +974,9 @@ impl Client {
         tx_args: TransactionArgs,
     ) -> Result<InputNotes<InputNote>, ClientError> {
         loop {
-            let execution = NoteConsumptionChecker::new(&self.tx_executor)
+            let data_store = ClientDataStore::new(self.store.clone());
+
+            let execution = NoteConsumptionChecker::new(&self.build_executor(&data_store)?)
                 .check_notes_consumability(
                     account_id,
                     self.store.get_sync_height().await?,
@@ -1139,15 +1120,16 @@ impl Client {
             .ok_or(ClientError::AccountDataNotFound(account_id))?;
         let account: Account = account_record.into();
 
+        let data_store = ClientDataStore::new(self.store.clone());
+
         // Ensure code is loaded on MAST store
-        self.mast_store.insert(tx_script.mast());
-        self.mast_store.insert(account.code().mast());
+        data_store.mast_store().load_account_code(account.code());
         for fpi_account in &foreign_account_inputs {
-            self.mast_store.insert(fpi_account.code().mast());
+            data_store.mast_store().load_account_code(fpi_account.code());
         }
 
         Ok(self
-            .tx_executor
+            .build_executor(&data_store)?
             .execute_tx_view_script(
                 account_id,
                 block_ref,
@@ -1157,6 +1139,17 @@ impl Client {
                 Arc::new(DefaultSourceManager::default()), // TODO: Use the correct source manager
             )
             .await?)
+    }
+
+    pub(crate) fn build_executor<'store, 'auth>(
+        &'auth self,
+        data_store: &'store ClientDataStore,
+    ) -> Result<TransactionExecutor<'store, 'auth>, TransactionExecutorError> {
+        TransactionExecutor::with_options(
+            data_store,
+            self.authenticator.as_deref(),
+            self.exec_options,
+        )
     }
 }
 
@@ -1229,6 +1222,33 @@ pub fn notes_from_output(output_notes: &OutputNotes) -> impl Iterator<Item = &No
         })
 }
 
+/// Validates that the executed transaction's output recipients match what was expected in the
+/// transaction request.
+fn validate_executed_transaction(
+    executed_transaction: &ExecutedTransaction,
+    expected_output_recipients: &[NoteRecipient],
+) -> Result<(), ClientError> {
+    let tx_output_recipient_digests = executed_transaction
+        .output_notes()
+        .iter()
+        .filter_map(|n| n.recipient().map(NoteRecipient::digest))
+        .collect::<Vec<_>>();
+
+    let missing_recipient_digest: Vec<Digest> = expected_output_recipients
+        .iter()
+        .filter_map(|recipient| {
+            (!tx_output_recipient_digests.contains(&recipient.digest()))
+                .then_some(recipient.digest())
+        })
+        .collect();
+
+    if !missing_recipient_digest.is_empty() {
+        return Err(ClientError::MissingOutputRecipients(missing_recipient_digest));
+    }
+
+    Ok(())
+}
+
 #[cfg(test)]
 mod test {
     use miden_lib::{account::auth::RpoFalcon512, transaction::TransactionKernel};
@@ -1255,7 +1275,7 @@ mod test {
     };
 
     #[tokio::test]
-    async fn test_transaction_creates_two_notes() {
+    async fn transaction_creates_two_notes() {
         let (mut client, _, keystore) = create_test_client().await;
         let asset_1: Asset =
             FungibleAsset::new(ACCOUNT_ID_PRIVATE_FUNGIBLE_FAUCET.try_into().unwrap(), 123)