+++ crates/web-client/test/fpi.test.ts
@@ -72,7 +72,6 @@ export const testStandardFpi = async (): Promise<void> => {
                     call.::miden::contracts::auth::basic::auth_tx_rpo_falcon512 
                 end
             `,
-      new window.TransactionScriptInputPairArray(),
       window.TransactionKernel.assembler()
     );
 
@@ -123,7 +122,6 @@ export const testStandardFpi = async (): Promise<void> => {
 
     let compiledTxScript = window.TransactionScript.compile(
       txScript,
-      new window.TransactionScriptInputPairArray(),
       window.TransactionKernel.assembler()
     );
 
@@ -152,7 +150,7 @@ export const testStandardFpi = async (): Promise<void> => {
 };
 
 describe("fpi test", () => {
-  it.only("runs the standard fpi test successfully", async () => {
+  it("runs the standard fpi test successfully", async () => {
     await expect(testStandardFpi()).to.be.fulfilled;
   }).timeout(50000);
 });