+++ tests/src/custom_transactions_tests.rs
@@ -48,7 +48,7 @@ const NOTE_ARGS: [Felt; 8] = [
 ];
 
 #[tokio::test]
-async fn test_transaction_request() {
+async fn transaction_request() {
     let (mut client, authenticator) = create_test_client().await;
     wait_for_node(&mut client).await;
 
@@ -93,7 +93,7 @@ async fn test_transaction_request() {
 
     let failure_code = code.replace("{asserted_value}", "1");
 
-    let tx_script = client.compile_tx_script(vec![], &failure_code).unwrap();
+    let tx_script = client.compile_tx_script(&failure_code).unwrap();
 
     let transaction_request = TransactionRequestBuilder::new()
         .with_authenticated_input_notes(note_args_map.clone())
@@ -109,7 +109,7 @@ async fn test_transaction_request() {
 
     let success_code = code.replace("{asserted_value}", "0");
 
-    let tx_script = client.compile_tx_script(vec![], &success_code).unwrap();
+    let tx_script = client.compile_tx_script(&success_code).unwrap();
 
     let transaction_request = TransactionRequestBuilder::new()
         .with_authenticated_input_notes(note_args_map)
@@ -131,7 +131,7 @@ async fn test_transaction_request() {
 }
 
 #[tokio::test]
-async fn test_merkle_store() {
+async fn merkle_store() {
     let (mut client, authenticator) = create_test_client().await;
     wait_for_node(&mut client).await;
 
@@ -206,7 +206,7 @@ async fn test_merkle_store() {
     code += "call.auth_tx::auth_tx_rpo_falcon512 end";
 
     // Build the transaction
-    let tx_script = client.compile_tx_script(vec![], &code).unwrap();
+    let tx_script = client.compile_tx_script(&code).unwrap();
 
     let transaction_request = TransactionRequestBuilder::new()
         .with_authenticated_input_notes(note_args_map)
@@ -222,7 +222,7 @@ async fn test_merkle_store() {
 }
 
 #[tokio::test]
-async fn test_onchain_notes_sync_with_tag() {
+async fn onchain_notes_sync_with_tag() {
     // Client 1 has an private faucet which will mint an onchain note for client 2
     let (mut client_1, keystore_1) = create_test_client().await;
     // Client 2 will be used to sync and check that by adding the tag we can still fetch notes
@@ -275,7 +275,7 @@ async fn test_onchain_notes_sync_with_tag() {
         .build()
         .unwrap();
 
-    let note = tx_request.expected_output_notes().next().unwrap().clone();
+    let note = tx_request.expected_output_own_notes().pop().unwrap().clone();
     execute_tx_and_sync(&mut client_1, basic_account_1.id(), tx_request).await;
 
     // Load tag into client 2