+++ tests/src/main_tests.rs
@@ -21,11 +21,12 @@ use winter_maybe_async::maybe_async_trait;
 
 mod custom_transactions_tests;
 mod fpi_tests;
+mod network_transaction_tests;
 mod onchain_tests;
 mod swap_transactions_tests;
 
 #[tokio::test]
-async fn test_client_builder_initializes_client_with_endpoint() -> Result<(), ClientError> {
+async fn client_builder_initializes_client_with_endpoint() -> Result<(), ClientError> {
     let (_, _, store_config, auth_path) = get_client_config();
 
     let mut client = ClientBuilder::new()
@@ -36,7 +37,7 @@ async fn test_client_builder_initializes_client_with_endpoint() -> Result<(), Cl
         .build()
         .await?;
 
-    assert!(client.is_in_debug_mode());
+    assert!(client.in_debug_mode());
 
     let sync_summary = client.sync_state().await.expect("Sync state failed");
 
@@ -46,32 +47,7 @@ async fn test_client_builder_initializes_client_with_endpoint() -> Result<(), Cl
 }
 
 #[tokio::test]
-async fn test_client_builder_initializes_client_with_rpc() -> Result<(), ClientError> {
-    let (_, _, store_config, auth_path) = get_client_config();
-
-    let endpoint = Endpoint::localhost();
-    let timeout_ms = 10_000;
-    let rpc_api = Arc::new(TonicRpcClient::new(&endpoint, timeout_ms));
-
-    let mut client = ClientBuilder::new()
-        .with_rpc(rpc_api)
-        .with_filesystem_keystore(auth_path.to_str().unwrap())
-        .with_sqlite_store(store_config.to_str().unwrap())
-        .in_debug_mode(true)
-        .build()
-        .await?;
-
-    assert!(client.is_in_debug_mode());
-
-    let sync_summary = client.sync_state().await.expect("Sync state failed");
-
-    assert!(sync_summary.block_num.as_u32() > 0);
-
-    Ok(())
-}
-
-#[tokio::test]
-async fn test_client_builder_fails_without_keystore() {
+async fn client_builder_fails_without_keystore() {
     let (_, _, store_config, _) = get_client_config();
     let result = ClientBuilder::new()
         .with_tonic_rpc_client(&Endpoint::default(), Some(10_000))
@@ -84,7 +60,7 @@ async fn test_client_builder_fails_without_keystore() {
 }
 
 #[tokio::test]
-async fn test_multiple_tx_on_same_block() {
+async fn multiple_tx_on_same_block() {
     let (mut client, authenticator) = create_test_client().await;
     wait_for_node(&mut client).await;
 
@@ -179,7 +155,7 @@ async fn test_multiple_tx_on_same_block() {
 }
 
 #[tokio::test]
-async fn test_import_expected_notes() {
+async fn import_expected_notes() {
     let (mut client_1, authenticator_1) = create_test_client().await;
     let (first_basic_account, faucet_account) =
         setup_wallet_and_faucet(&mut client_1, AccountStorageMode::Private, &authenticator_1).await;
@@ -200,7 +176,8 @@ async fn test_import_expected_notes() {
             client_2.rng(),
         )
         .unwrap();
-    let note: InputNoteRecord = tx_request.expected_output_notes().next().unwrap().clone().into();
+    let note: InputNoteRecord =
+        tx_request.expected_output_own_notes().pop().unwrap().clone().into();
     client_2.sync_state().await.unwrap();
 
     // If the verification is requested before execution then the import should fail
@@ -234,7 +211,8 @@ async fn test_import_expected_notes() {
             client_2.rng(),
         )
         .unwrap();
-    let note: InputNoteRecord = tx_request.expected_output_notes().next().unwrap().clone().into();
+    let note: InputNoteRecord =
+        tx_request.expected_output_own_notes().pop().unwrap().clone().into();
 
     // Import an uncommited note without verification
     client_2.add_note_tag(note.metadata().unwrap().tag()).await.unwrap();
@@ -264,7 +242,7 @@ async fn test_import_expected_notes() {
 }
 
 #[tokio::test]
-async fn test_import_expected_note_uncommitted() {
+async fn import_expected_note_uncommitted() {
     let (mut client_1, authenticator) = create_test_client().await;
     let faucet_account =
         insert_new_fungible_faucet(&mut client_1, AccountStorageMode::Private, &authenticator)
@@ -289,7 +267,8 @@ async fn test_import_expected_note_uncommitted() {
         )
         .unwrap();
 
-    let note: InputNoteRecord = tx_request.expected_output_notes().next().unwrap().clone().into();
+    let note: InputNoteRecord =
+        tx_request.expected_output_own_notes().pop().unwrap().clone().into();
     client_2.sync_state().await.unwrap();
 
     // If the verification is requested before execution then the import should fail
@@ -308,7 +287,7 @@ async fn test_import_expected_note_uncommitted() {
 }
 
 #[tokio::test]
-async fn test_import_expected_notes_from_the_past_as_committed() {
+async fn import_expected_notes_from_the_past_as_committed() {
     let (mut client_1, authenticator_1) = create_test_client().await;
     let (first_basic_account, faucet_account) =
         setup_wallet_and_faucet(&mut client_1, AccountStorageMode::Private, &authenticator_1).await;
@@ -325,7 +304,8 @@ async fn test_import_expected_notes_from_the_past_as_committed() {
             client_1.rng(),
         )
         .unwrap();
-    let note: InputNoteRecord = tx_request.expected_output_notes().next().unwrap().clone().into();
+    let note: InputNoteRecord =
+        tx_request.expected_output_own_notes().pop().unwrap().clone().into();
 
     let block_height_before = client_1.get_sync_height().await.unwrap();
 
@@ -354,7 +334,7 @@ async fn test_import_expected_notes_from_the_past_as_committed() {
 }
 
 #[tokio::test]
-async fn test_get_account_update() {
+async fn get_account_update() {
     // Create a client with both public and private accounts.
     let (mut client, authenticator) = create_test_client().await;
 
@@ -386,7 +366,7 @@ async fn test_get_account_update() {
 }
 
 #[tokio::test]
-async fn test_sync_detail_values() {
+async fn sync_detail_values() {
     let (mut client1, authenticator_1) = create_test_client().await;
     let (mut client2, authenticator_2) = create_test_client().await;
     wait_for_node(&mut client1).await;
@@ -425,7 +405,7 @@ async fn test_sync_detail_values() {
             client1.rng(),
         )
         .unwrap();
-    let note_id = tx_request.expected_output_notes().next().unwrap().id();
+    let note_id = tx_request.expected_output_own_notes().pop().unwrap().id();
     execute_tx_and_sync(&mut client1, from_account_id, tx_request).await;
 
     // Second client sync should have new note
@@ -448,7 +428,7 @@ async fn test_sync_detail_values() {
 /// This test runs 3 mint transactions that get included in different blocks so that once we sync
 /// we can check that each transaction gets marked as committed in the corresponding block.
 #[tokio::test]
-async fn test_multiple_transactions_can_be_committed_in_different_blocks_without_sync() {
+async fn multiple_transactions_can_be_committed_in_different_blocks_without_sync() {
     let (mut client, authenticator) = create_test_client().await;
 
     let (first_regular_account, faucet_account_header) =
@@ -478,7 +458,7 @@ async fn test_multiple_transactions_can_be_committed_in_different_blocks_without
         let transaction_id = transaction_execution_result.executed_transaction().id();
 
         println!("Sending transaction to node");
-        let note_id = tx_request.expected_output_notes().next().unwrap().id();
+        let note_id = tx_request.expected_output_own_notes().pop().unwrap().id();
         client.submit_transaction(transaction_execution_result).await.unwrap();
 
         (note_id, transaction_id)
@@ -506,7 +486,7 @@ async fn test_multiple_transactions_can_be_committed_in_different_blocks_without
 
         println!("Sending transaction to node");
         // May need a few attempts until it gets included
-        let note_id = tx_request.expected_output_notes().next().unwrap().id();
+        let note_id = tx_request.expected_output_own_notes().pop().unwrap().id();
         while client
             .test_rpc_api()
             .get_notes_by_id(&[first_note_id])
@@ -543,7 +523,7 @@ async fn test_multiple_transactions_can_be_committed_in_different_blocks_without
 
         println!("Sending transaction to node");
         // May need a few attempts until it gets included
-        let note_id = tx_request.expected_output_notes().next().unwrap().id();
+        let note_id = tx_request.expected_output_own_notes().pop().unwrap().id();
         while client
             .test_rpc_api()
             .get_notes_by_id(&[second_note_id])
@@ -596,7 +576,7 @@ async fn test_multiple_transactions_can_be_committed_in_different_blocks_without
 /// - Consuming authenticated notes.
 /// - Consuming unauthenticated notes.
 #[tokio::test]
-async fn test_consume_multiple_expected_notes() {
+async fn consume_multiple_expected_notes() {
     let (mut client, authenticator_1) = create_test_client().await;
     let (mut unauth_client, authenticator_2) = create_test_client().await;
 
@@ -627,7 +607,7 @@ async fn test_consume_multiple_expected_notes() {
     unauth_client.sync_state().await.unwrap();
 
     // Filter notes by ownership
-    let expected_notes = mint_tx_request.expected_output_notes();
+    let expected_notes = mint_tx_request.expected_output_own_notes().into_iter();
     let client_notes: Vec<_> = client.get_input_notes(NoteFilter::All).await.unwrap();
     let client_notes_ids: Vec<_> = client_notes.iter().map(|note| note.id()).collect();
 
@@ -685,7 +665,7 @@ async fn test_consume_multiple_expected_notes() {
 }
 
 #[tokio::test]
-async fn test_import_consumed_note_with_proof() {
+async fn import_consumed_note_with_proof() {
     let (mut client_1, authenticator_1) = create_test_client().await;
     let (first_regular_account, faucet_account_header) =
         setup_wallet_and_faucet(&mut client_1, AccountStorageMode::Private, &authenticator_1).await;
@@ -749,7 +729,7 @@ async fn test_import_consumed_note_with_proof() {
 }
 
 #[tokio::test]
-async fn test_import_consumed_note_with_id() {
+async fn import_consumed_note_with_id() {
     let (mut client_1, authenticator) = create_test_client().await;
     let (first_regular_account, second_regular_account, faucet_account_header) =
         setup_two_wallets_and_faucet(&mut client_1, AccountStorageMode::Private, &authenticator)
@@ -805,7 +785,7 @@ async fn test_import_consumed_note_with_id() {
 }
 
 #[tokio::test]
-async fn test_import_note_with_proof() {
+async fn import_note_with_proof() {
     let (mut client_1, authenticator) = create_test_client().await;
     let (first_regular_account, second_regular_account, faucet_account_header) =
         setup_two_wallets_and_faucet(&mut client_1, AccountStorageMode::Private, &authenticator)
@@ -868,7 +848,7 @@ async fn test_import_note_with_proof() {
 }
 
 #[tokio::test]
-async fn test_discarded_transaction() {
+async fn discarded_transaction() {
     let (mut client_1, authenticator_1) = create_test_client().await;
     let (first_regular_account, faucet_account_header) =
         setup_wallet_and_faucet(&mut client_1, AccountStorageMode::Private, &authenticator_1).await;
@@ -997,7 +977,7 @@ impl TransactionProver for AlwaysFailingProver {
 }
 
 #[tokio::test]
-async fn test_custom_transaction_prover() {
+async fn custom_transaction_prover() {
     let (mut client, authenticator) = create_test_client().await;
     let (first_regular_account, faucet_account_header) =
         setup_wallet_and_faucet(&mut client, AccountStorageMode::Private, &authenticator).await;
@@ -1031,7 +1011,7 @@ async fn test_custom_transaction_prover() {
 }
 
 #[tokio::test]
-async fn test_locked_account() {
+async fn locked_account() {
     let (mut client_1, authenticator) = create_test_client().await;
 
     let (faucet_account, ..) =
@@ -1084,7 +1064,7 @@ async fn test_locked_account() {
 }
 
 #[tokio::test]
-async fn test_expired_transaction_fails() {
+async fn expired_transaction_fails() {
     let (mut client, authenticator) = create_test_client().await;
     let (faucet_account, ..) =
         insert_new_fungible_faucet(&mut client, AccountStorageMode::Private, &authenticator)
@@ -1127,7 +1107,7 @@ async fn test_expired_transaction_fails() {
 /// Tests that RPC methods that are not directly related to the client logic
 /// (like GetBlockByNumber) work correctly
 #[tokio::test]
-async fn test_unused_rpc_api() {
+async fn unused_rpc_api() {
     let (mut client, keystore) = create_test_client().await;
 
     let (first_basic_account, faucet_account) =
@@ -1188,7 +1168,7 @@ async fn test_unused_rpc_api() {
 }
 
 #[tokio::test]
-async fn test_ignore_invalid_notes() {
+async fn ignore_invalid_notes() {
     let (mut client, authenticator) = create_test_client().await;
     let (regular_account, second_regular_account, faucet_account_header) =
         setup_two_wallets_and_faucet(&mut client, AccountStorageMode::Private, &authenticator)